requirements.txt
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-nested-admin==3.3.3
djangorestframework==3.12.4
gunicorn==20.1.0
python-monkey-business==1.0.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
whitenoise==5.2.0
